# contracts/openapi-platform.yaml
openapi: 3.1.0
info:
  title: Enterprise Data Copilot Platform API
  version: 0.1.0
  description: Browser-facing API owned by the Java platform service.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      operationId: getPlatformHealth
      tags: [operations]
      security: []
      responses:
        "200":
          description: The platform process is running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/auth/login:
    post:
      operationId: login
      tags: [identity]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: A short-lived platform access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/data-sources:
    get:
      operationId: listDataSources
      tags: [data-sources]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Data sources in the authenticated tenant.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataSourceResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createDataSource
      description: ADMIN only. Credentials are supplied by reference and are never returned.
      tags: [data-sources]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDataSourceRequest"
      responses:
        "201":
          description: Tenant-scoped data source metadata was registered.
          headers:
            Location:
              required: true
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataSourceResource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/analysis/jobs:
    get:
      operationId: listAnalysisJobs
      description: Returns at most the 50 most recent jobs visible to the caller.
      tags: [analysis]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tenant-scoped analysis history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalysisJobResource"
    post:
      operationId: createAnalysisJob
      tags: [analysis]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisJobRequest"
      responses:
        "202":
          description: The analysis job was accepted.
          headers:
            Location:
              description: URL of the accepted analysis job.
              required: true
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJobResource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/analysis/jobs/{jobId}:
    get:
      operationId: getAnalysisJob
      tags: [analysis]
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current analysis job state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisJobResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/analysis/jobs/{jobId}/steps:
    get:
      operationId: listAgentSteps
      tags: [analysis]
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Redacted, tenant-scoped workflow step summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentStepResource"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/analysis/jobs/{jobId}/events:
    get:
      operationId: streamAnalysisJobEvents
      description: Replays retained events after Last-Event-ID and then streams live progress.
      tags: [analysis]
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ordered, tenant-scoped analysis progress events.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/AnalysisJobEvent"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/approvals/{approvalId}/approve:
    post:
      operationId: approveAnalysisRun
      tags: [approvals]
      security:
        - bearerAuth: []
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: The administrator decision and resumed workflow state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResource"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/approvals:
    get:
      operationId: listApprovals
      description: ADMIN-only tenant approval history, newest first.
      tags: [approvals]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tenant-scoped approvals.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalResource"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/approvals/{approvalId}/reject:
    post:
      operationId: rejectAnalysisRun
      tags: [approvals]
      security:
        - bearerAuth: []
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: The administrator rejection and terminal job state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResource"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/metric-documents:
    get:
      operationId: listMetricDocuments
      tags: [knowledge]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Metric document versions in the authenticated tenant.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetricDocumentResource"
    post:
      operationId: uploadMetricDocument
      description: ADMIN-only bounded document ingestion through the internal Agent service.
      tags: [knowledge]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetricDocumentUploadRequest"
      responses:
        "201":
          description: The document was parsed, versioned and indexed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricDocumentResource"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key for safely replaying a create request.
      schema:
        type: string
        minLength: 16
        maxLength: 128
  responses:
    BadRequest:
      description: The request is malformed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: The caller cannot access the requested tenant resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The resource was not found in the caller's authorized scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: The idempotency key conflicts with an earlier request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    UnprocessableEntity:
      description: The request is well formed but violates an input rule.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InternalError:
      description: The platform could not complete the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required: [service, status, timestamp]
      properties:
        service:
          type: string
          const: platform-api
        status:
          type: string
          const: UP
        timestamp:
          type: string
          format: date-time
    AnalysisJobStatus:
      type: string
      enum:
        - CREATED
        - PLANNING
        - VALIDATING
        - WAITING_APPROVAL
        - RUNNING
        - COMPLETED
        - REJECTED
        - FAILED
        - CANCELLED
    CreateAnalysisJobRequest:
      type: object
      additionalProperties: false
      required: [data_source_id, question]
      properties:
        data_source_id:
          type: string
          format: uuid
        question:
          type: string
          minLength: 1
          maxLength: 4000
    AnalysisJobResource:
      type: object
      additionalProperties: false
      required:
        - id
        - data_source_id
        - question
        - status
        - trace_id
        - created_at
        - updated_at
        - version
        - columns
        - rows
        - chart
        - citations
      properties:
        id:
          type: string
          format: uuid
        data_source_id:
          type: string
          format: uuid
        question:
          type: string
        status:
          $ref: "#/components/schemas/AnalysisJobStatus"
        trace_id:
          type: string
          minLength: 16
          maxLength: 64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        error:
          $ref: "#/components/schemas/ApiError"
        version:
          type: integer
          format: int64
          minimum: 0
        workflow_thread_id:
          type: string
          maxLength: 64
        generated_sql:
          type: string
        answer:
          type: string
        error_code:
          type: string
          maxLength: 80
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: true
        chart:
          $ref: "#/components/schemas/ChartSpec"
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
    ChartSpec:
      type: object
      additionalProperties: false
      required: [type, title, series]
      properties:
        type:
          type: string
          enum: [bar, line, pie, table]
        title:
          type: string
          maxLength: 160
        x:
          type: [string, "null"]
        series:
          type: array
          maxItems: 3
          items:
            type: string
    Citation:
      type: object
      additionalProperties: true
      properties:
        document_id:
          type: string
          format: uuid
        document_title:
          type: string
        document_version:
          type: integer
        section_key:
          type: string
        section_title:
          type: string
        source_locator:
          type: string
    AnalysisJobEvent:
      type: object
      additionalProperties: false
      required:
        [event_version, event_id, sequence, job_id, tenant_id, trace_id, type, status, occurred_at, payload]
      properties:
        event_version:
          type: string
          const: "1.0"
        event_id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
          minimum: 0
        job_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        trace_id:
          type: string
          minLength: 16
          maxLength: 64
        type:
          type: string
          enum:
            [JOB_ACCEPTED, JOB_RECOVERED, JOB_STATUS_CHANGED, JOB_SNAPSHOT, AGENT_STEP, APPROVAL_REQUIRED, APPROVAL_DECIDED]
        status:
          $ref: "#/components/schemas/AnalysisJobStatus"
        occurred_at:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    AgentStepResource:
      type: object
      additionalProperties: false
      required:
        [id, step_name, status, attempt, input_summary, output_summary, duration_ms, created_at]
      properties:
        id:
          type: string
          format: uuid
        step_name:
          type: string
          maxLength: 80
        status:
          type: string
          enum: [SUCCEEDED, FAILED, WAITING]
        attempt:
          type: integer
          minimum: 0
          maximum: 3
        input_summary:
          type: object
          additionalProperties: true
        output_summary:
          type: object
          additionalProperties: true
        duration_ms:
          type: integer
          format: int64
          minimum: 0
        error_code:
          type: string
          maxLength: 80
        created_at:
          type: string
          format: date-time
    ApprovalDecisionRequest:
      type: object
      additionalProperties: false
      properties:
        comment:
          type: string
          maxLength: 500
    ApprovalResource:
      type: object
      additionalProperties: false
      required: [id, analysis_job_id, status, reason, created_at, version]
      properties:
        id:
          type: string
          format: uuid
        analysis_job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        reason:
          type: string
          maxLength: 500
        decided_by:
          type: string
          format: uuid
        decision_comment:
          type: string
          maxLength: 500
        created_at:
          type: string
          format: date-time
        decided_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 0
    LoginRequest:
      type: object
      additionalProperties: false
      required: [tenant_slug, email, password]
      properties:
        tenant_slug:
          type: string
          minLength: 1
          maxLength: 80
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 200
    TokenResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in, user_id, tenant_id, roles]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          minimum: 1
        user_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [ADMIN, ANALYST, VIEWER]
    CreateDataSourceRequest:
      type: object
      additionalProperties: false
      required: [name, host, port, database_name, allowed_schema, secret_ref]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        host:
          type: string
          minLength: 1
          maxLength: 255
        port:
          type: integer
          minimum: 1
          maximum: 65535
        database_name:
          type: string
          minLength: 1
          maxLength: 128
        allowed_schema:
          type: string
          pattern: "^[a-z_][a-z0-9_]*$"
          maxLength: 63
        secret_ref:
          type: string
          writeOnly: true
          pattern: "^(env|vault|secret):[A-Za-z0-9_./-]+$"
          maxLength: 255
    DataSourceResource:
      type: object
      additionalProperties: false
      required: [id, name, source_type, allowed_schema, enabled, version]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        source_type:
          type: string
          const: POSTGRESQL
        allowed_schema:
          type: string
        enabled:
          type: boolean
        version:
          type: integer
          format: int64
          minimum: 0
    MetricDocumentUploadRequest:
      type: object
      additionalProperties: false
      required: [title, source_name, source_type, content_base64]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 240
        source_name:
          type: string
          minLength: 1
          maxLength: 255
        source_type:
          type: string
          enum: [MARKDOWN, PDF]
        content_base64:
          type: string
          contentEncoding: base64
          maxLength: 14000000
    MetricDocumentResource:
      type: object
      additionalProperties: false
      required: [id, tenant_id, title, version, status, source_type, source_name, content_sha256, chunk_count, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        title:
          type: string
        version:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [ACTIVE, SUPERSEDED]
        source_type:
          type: string
          enum: [MARKDOWN, PDF]
        source_name:
          type: string
        content_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        chunk_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, trace_id, timestamp]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 80
        message:
          type: string
          minLength: 1
          maxLength: 500
        trace_id:
          type: string
          minLength: 16
          maxLength: 64
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          additionalProperties: true